+++ src/App.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect } from "react";
 import { Routes, Route } from "react-router-dom";
-import { Navigation, Home, ProjectsList } from "./pages";
+import { Navigation, Home, ProjectsList, ProjectDetails } from "./pages";
 import ScrollToTopButton from "./components/ScrollToTopButton";
+import { projects } from "./utils/constants";
 
 function App() {
   const [isDarkMode, setDarkMode] = useState(false);
@@ -58,6 +59,19 @@ function App() {
           path="/project/all"
           element={<ProjectsList isDarkMode={isDarkMode} isMobile={isMobile} />}
         />
+        {projects.map((project, index) => (
+          <Route
+            path={`/project/${project.id}`}
+            element={
+              <ProjectDetails
+                isDarkMode={isDarkMode}
+                isMobile={isMobile}
+                project={project}
+                key={index}
+              />
+            }
+          />
+        ))}
       </Routes>
       <ScrollToTopButton isMobile={isMobile} />
     </div>