+++ src/assets/index.ts
@@ -1,10 +1,13 @@
 import creditsafe from "./company_logos/creditsafe.png";
 import financialforce from "./company_logos/financialforce.png";
-import { instock } from "./project_screenshots/instock";
-import { wrackreport } from "./project_screenshots/wrackreport";
-import { prayerpal } from "./project_screenshots/prayerpal";
-import { e2s } from "./project_screenshots/e2s";
-import { hafod } from "./project_screenshots/hafod";
+import { instock, instockDemo } from "./project_screenshots/instock";
+import {
+  wrackreport,
+  wrackreportDemo,
+} from "./project_screenshots/wrackreport";
+import { prayerpal, prayerpalDemo } from "./project_screenshots/prayerpal";
+import { e2s, e2sDemo } from "./project_screenshots/e2s";
+import { hafod, hafodDemo } from "./project_screenshots/hafod";
 import wave from "./loading/wave.png";
 
 export {
@@ -17,3 +20,5 @@ export {
   hafod,
   wave,
 };
+
+export { instockDemo, wrackreportDemo, prayerpalDemo, e2sDemo, hafodDemo };