+++ src/assets/project_screenshots/hafod/index.ts
@@ -5,13 +5,16 @@ import footer from "./footer.png";
 import tenants from "./tenants.png";
 import tenantSearch from "./tenant_search.png";
 import addTenant from "./add_tenant.png";
+import demo from "./hafod_demo.mp4";
 
 export const hafod = [
-    login,
-    home,
-    charts,
-    footer,
-    tenants,
-    tenantSearch,
-    addTenant,
-]
+  login,
+  home,
+  charts,
+  footer,
+  tenants,
+  tenantSearch,
+  addTenant,
+];
+
+export const hafodDemo = demo;