+++ src/assets/project_screenshots/prayerpal/index.ts
@@ -5,13 +5,16 @@ import tracker from "./tracker.jpeg";
 import duas from "./duas.jpeg";
 import duaSearch from "./dua_search.jpeg";
 import dua from "./dua.jpeg";
+import demo from "./prayerpal_demo.mp4";
 
 export const prayerpal = [
-    prayerTimes,
-    map,
-    qibla,
-    tracker,
-    duas,
-    duaSearch,
-    dua,
-]
+  prayerTimes,
+  map,
+  qibla,
+  tracker,
+  duas,
+  duaSearch,
+  dua,
+];
+
+export const prayerpalDemo = demo;