+++ src/pages/index.ts
@@ -5,5 +5,15 @@ import About from "./about/About";
 import Experiences from "./experience/Experiences";
 import Projects from "./projects/Projects";
 import ProjectsList from "./projects/ProjectsList";
+import ProjectDetails from "./projects/ProjectDetails";
 
-export { Navigation, Home, Landing, About, Experiences, Projects, ProjectsList };
+export {
+  Navigation,
+  Home,
+  Landing,
+  About,
+  Experiences,
+  Projects,
+  ProjectsList,
+  ProjectDetails,
+};